rename cypress command and change the options
