+++ projects/storefrontapp-e2e-cypress/cypress/e2e/b2b/regression/user-registration/b2b-otp-registration-e2e-spec-flaky.cy.ts
@@ -68,7 +68,7 @@ describe('Tabbing order for B2B OTP registration', () => {
           cy.get('cx-global-message').contains(
             'Thank you for registering! A representative will contact you shortly and confirm your access information.'
           );
-          cy.loginRegisterLinkSelect(false).should('exist');
+          cy.getLoginRegisterLink().should('exist');
         });
       });
     });
