+++ projects/storefrontapp-e2e-cypress/cypress/e2e/regression/anonymous-consents/anonymous-consents-config-flow.e2e.cy.ts
@@ -46,7 +46,7 @@ context('Anonymous consents - config flow', () => {
         cy.wait(`@${homePage}`).its('response.statusCode').should('eq', 200);
 
         // Make sure anonymous user is loaded
-        cy.loginRegisterLinkSelect(false).should('be.visible');
+        cy.getLoginRegisterLink().should('be.visible');
         seeBannerAsAnonymous();
 
         // Make sure user is logged in after saving it in storage
