+++ projects/storefrontapp-e2e-cypress/cypress/e2e/regression/user_access/mock-captcha.core-e2e.cy.ts
@@ -32,7 +32,7 @@ describe('Register', () => {
       cy.onMobile(() => {
         clickHamburger();
       });
-      cy.loginRegisterLinkSelect(true);
+      cy.getLoginRegisterLink({ clickAndWait: true });
       cy.get('cx-login-register').findByText('Register').click();
       cy.get('cx-captcha').should('exist');
       registerWithCaptcha(user);
