+++ projects/storefrontapp-e2e-cypress/cypress/helpers/b2b/b2b-user-registration.ts
@@ -28,7 +28,7 @@ export function navigateToOrganizationUserRegisterPage() {
   cy.onMobile(() => {
     clickHamburger();
   });
-  cy.loginRegisterLinkSelect(true);
+  cy.getLoginRegisterLink({ clickAndWait: true });
   cy.get(ORGANIZATION_USER_REGISTER_BUTTON_SELECTOR).find('a').click();
 }
