+++ projects/storefrontapp-e2e-cypress/cypress/helpers/order-history.ts
@@ -101,7 +101,7 @@ export const orderHistoryTest = {
     it('should redirect to login page if user is not logged in', () => {
       cy.visit(url);
       cy.url().should('contain', '/login');
-      cy.loginRegisterLinkSelect(false).should('contain', 'Sign In / Register');
+      cy.getLoginRegisterLink().should('contain', 'Sign In / Register');
     });
   },
   checkRedirectLoggedInUser(
